requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
MySQL-connector-python==8.1.0
python-dotenv==1.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow==10.0.0
email-validator==2.0.0
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0 